Removed warnings

+++ src/samara.h
@@ -11,7 +11,7 @@ pair <vector <string>, vector < vector <double> > > run_samara_2_1_micha(SamaraP
 pair <vector <string>, vector < vector <double> > > run_samara_2_3(SamaraParameters * parameters, SamaraLogType log = SMALL);
 pair <vector <string>, vector < vector <double> > > run_samara_2_3_lodging(SamaraParameters * parameters, SamaraLogType log = SMALL);
 
-void set_meteo_vars(SamaraParameters * parameters, double t, double &TMax, double &TMin, double &TMoy
+void set_meteo_vars(SamaraParameters * parameters, int t, double &TMax, double &TMin, double &TMoy
                     , double &HMax, double &HMin, double &HMoy
                     , double &Vt, double &Ins, double &Rg, double &ETP
                     , double &Pluie, double & TMoyCalc, double & HMoyCalc);
